moved frobenius_polynomial, fixed a bug in frobpoly, optimized finding ap in frobpoly
